+++ src/app/components/vc-view/vc-view.component.ts
@@ -15,9 +15,6 @@ import {
   VerifiableCredential,
 } from 'src/app/interfaces/verifiable-credential';
 import { IonicModule } from '@ionic/angular';
-import { HttpErrorResponse, HttpResponse } from '@angular/common/http';
-import { ToastServiceHandler } from 'src/app/services/toast.service';
-import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-vc-view',
@@ -80,10 +77,6 @@ export class VcViewComponent implements OnInit {
   },
   ];
   private walletService = inject(WalletService);
-  public constructor(
-    private toastServiceHandler: ToastServiceHandler,
-    private router: Router
-  ) {}
 
   public ngOnInit(): void {
     this.checkExpirationVC();
@@ -155,40 +148,11 @@ export class VcViewComponent implements OnInit {
     );
   }
 
-  /*public requestSignature(): void {
-    if (this.credentialInput?.id) {
-      this.walletService.requestSignature(this.credentialInput.id).subscribe({
-        next: (response: HttpResponse<string>) => {
-          if (response.status === 204) {
-            console.log(
-              'Credential request completed successfully, no content returned.'
-            );
-
-            this.toastServiceHandler.showErrorAlert('Unsigned').subscribe();
-          } else {
-            this.forcePageReload();
-          }
-        },
-        error: (error: HttpErrorResponse) => {
-          console.error('Error requesting signature:', error.message);
-          this.toastServiceHandler.showErrorAlert('ErrorUnsigned').subscribe();
-        },
-      });
-    }
-  }
-  public forcePageReload(): void {
-    this.router.navigate(['/tabs/credentials']).then(() => {
-      window.location.reload();
-    });
-  } */
-
   public handleKeydown(event: KeyboardEvent, action = 'request') {
     if (event.key === 'Enter' || event.key === ' ') {
       if (action === 'qr') {
         this.qrView();
-      } else {
-        //this.requestSignature();
-      }
+      } 
       event.preventDefault();
     }
   }
@@ -200,10 +164,9 @@ export class VcViewComponent implements OnInit {
       } else if (action === 'close') {
         this.setOpen(false);
       } else if (action === 'info') {
-        console.log('Action info');
         this.unsignedInfo();
       }
       event.preventDefault();
-  }
+    }
   }
 }